+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## New Features in This Release
+- feat(Application): support empty separator
 - feat(Application): more flexible custom separator
 - feat(Application): add exception alert
 - feat(Application): add printing sum